+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Version 1.5.1 (not released yet)
 - use dynamic allocation when pool is full for weapons, debris, corpses, decal polygons, faces, face vertices,
   bounding boxes, vertices, meshes (it was listed as a feature of Dash Faction 1.5.0 but the patch was actually
   broken)
+- support starting the launcher from a symlink
 
 Version 1.5.0 (released 2020-09-05)
 -----------------------------------